backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
anthropic==0.40.0
openai==1.57.0
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.2.5
Pillow==11.0.0
msal==1.31.0
httpx==0.28.0
python-dotenv==1.0.1
aiofiles==24.1.0
email-validator==2.2.0
